+++ CHANGELOG.md
@@ -25,6 +25,7 @@ commands, which is compatible with `sensuctl create`.
 `backend.yml.example`.
 - Added environment variables to checks.
 - Added logging redaction integration test.
+- Added check token substitution integration test.
 
 ### Changed
 - Add logging around the Sensu event pipeline.
@@ -86,6 +87,7 @@ lockfiles to not be cleaned up.
 integration coverage.
 - Removed the `leader` package
 - Removed logging redaction e2e test, in favor of integration coverage.
+- Removed check token substitution e2e test, in favor of integration coverage.
 
 ## [2.0.0-beta.1] - 2018-05-07
 ### Added